Split the server-actions file with all the executions logic into three files with more distinct responsibilities, namely deployments, instances and tasks; renamed the file in which we define the Engine object types
