+++ src/management-system-v2/app/(dashboard)/[environmentId]/(automation)/engines/[dbEngineId]/page.tsx
@@ -4,7 +4,7 @@ import { Button, Skeleton } from 'antd';
 import { notFound } from 'next/navigation';
 import { Suspense } from 'react';
 import { LeftOutlined } from '@ant-design/icons';
-import { type Engine } from '@/lib/engines/machines';
+import { type Engine } from '@/lib/engines/types';
 import { getDbEngineById } from '@/lib/data/engines';
 import { getMSConfig } from '@/lib/ms-config/ms-config';
 import EngineDashboard from '@/components/engine-dashboard/server-component';
